+++ debug_toolbar/decorators.py
@@ -1,6 +1,9 @@
 import functools
 
 from django.http import Http404
+from django.utils.translation import get_language, override as language_override
+
+from debug_toolbar import settings as dt_settings
 
 
 def require_show_toolbar(view):
@@ -13,5 +16,18 @@ def require_show_toolbar(view):
             raise Http404
 
         return view(request, *args, **kwargs)
+
+    return inner
+
+
+def render_with_toolbar_language(view):
+    """Force any rendering within the view to use the toolbar's language."""
+
+    @functools.wraps(view)
+    def inner(request, *args, **kwargs):
+
+        lang = dt_settings.get_config()["TOOLBAR_LANGUAGE"] or get_language()
+        with language_override(lang):
+            return view(request, *args, **kwargs)
 
     return inner