+++ debug_toolbar/panels/templates/views.py
@@ -5,10 +5,11 @@ from django.template.engine import Engine
 from django.template.loader import render_to_string
 from django.utils.html import format_html, mark_safe
 
-from debug_toolbar.decorators import require_show_toolbar
+from debug_toolbar.decorators import render_with_toolbar_language, require_show_toolbar
 
 
 @require_show_toolbar
+@render_with_toolbar_language
 def template_source(request):
     """
     Return the source of a template, syntax-highlighted by Pygments if