+++ debug_toolbar/views.py
@@ -2,11 +2,12 @@ from django.http import JsonResponse
 from django.utils.html import escape
 from django.utils.translation import gettext as _
 
-from debug_toolbar.decorators import require_show_toolbar
+from debug_toolbar.decorators import render_with_toolbar_language, require_show_toolbar
 from debug_toolbar.toolbar import DebugToolbar
 
 
 @require_show_toolbar
+@render_with_toolbar_language
 def render_panel(request):
     """Render the contents of a panel"""
     toolbar = DebugToolbar.fetch(request.GET["store_id"])